# src/main/resources/openapi/card-service.yaml
openapi: 3.0.3
info:
  title: Card Service API
  version: 1.0.0
  description: API for managing debit & credit cards
servers:
  - url: /api/v1
paths:
...
  /debit-cards:
    get:
      summary: Get all debit cards
      operationId: getAllDebitCards
      parameters:
        - name: customerId
          in: query
          required: false
          schema:
            type: string
          description: Filter debit cards by customer ID
        - name: isActive
          in: query
          required: false
          schema:
            type: boolean
          description: Filter debit cards by active status (default = true)
      responses:
        '200':
          description: List of debit cards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DebitCardResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create a new debit card
      operationId: createDebitCard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DebitCardCreateRequest'
      responses:
        '201':
          description: Debit card created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebitCardResponse'
        '400':
          description: Bad request - Validation errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Business rule violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /debit-cards/{id}:
    get:
      summary: Get a debit card by ID
      operationId: getDebitCardById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Debit card found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebitCardResponse'
        '404':
          description: Debit card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Update a debit card by ID
      operationId: updateDebitCard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DebitCardUpdateRequest'
      responses:
        '200':
          description: Debit card updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebitCardResponse'
        '404':
          description: Debit card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad request - Validation errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete a debit card by ID
      operationId: deleteDebitCard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Debit card deleted
        '404':
          description: Debit card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /debit-cards/{id}/deactivate:
    patch:
      summary: Deactivate a debit card by ID
      operationId: deactivateDebitCard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Debit card deactivated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebitCardResponse'
        '404':
          description: Debit card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /debit-cards/{id}/activate:
    patch:
      summary: Activate a debit card by ID
      operationId: activateDebitCard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Debit card activated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebitCardResponse'
        '404':
          description: Debit card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # DEBIT CARD SPECIFIC OPERATIONS
  /debit-cards/{id}/associate-account:
    post:
      summary: Associate an account to debit card
      description: Associates a bank account to the debit card. First account becomes primary if none exists.
      operationId: associateAccountToDebitCard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssociateAccountRequest'
      responses:
        '200':
          description: Account associated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebitCardResponse'
        '400':
          description: Invalid account or already associated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Debit card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /debit-cards/{cardNumber}/process-purchase:
    post:
      summary: Process debit card purchase/withdrawal
      description: Processes purchase using cascade logic across associated accounts
      operationId: processDebitCardPurchase
      parameters:
        - name: cardNumber
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DebitPurchaseRequest'
      responses:
        '200':
          description: Purchase processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebitPurchaseResponse'
        '400':
          description: Insufficient funds or invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Debit card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /debit-cards/{cardId}/primary-account/balance:
    get:
      summary: Get primary account balance via debit card
      description: Retrieves the balance of the primary account associated with the debit card
      operationId: getDebitCardPrimaryAccountBalance
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Primary account balance retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrimaryAccountBalanceResponse'
        '404':
          description: Debit card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    DebitCardCreateRequest:
      type: object
      properties:
        customerId:
          type: string
          example: "68a0cdde7d68c910da0bc239"
        primaryAccountId:
          type: string
          example: "acc_123456789"
        associatedAccountIds:
          type: array
          items:
            type: string
          example: [ "acc_987654321", "acc_456789123" ]
          description: Additional accounts in priority order
      required:
        - customerId
        - primaryAccountId
    DebitCardUpdateRequest:
      type: object
      properties:
        isActive:
          type: boolean
          example: true
        primaryAccountId:
          type: string
          example: "acc_123456789"
        associatedAccountIds:
          type: array
          items:
            type: string
          example: [ "acc_987654321", "acc_456789123" ]
      required:
        - isActive
    DebitCardResponse:
      type: object
      properties:
        id:
          type: string
          example: "64f3cbb5f8a1e53d2f7c34e2"
        cardNumber:
          type: string
          example: "1234-5678-9123-4568"
        customerId:
          type: string
          example: "68a0cdde7d68c910da0bc239"
        primaryAccountId:
          type: string
          example: "acc_123456789"
        associatedAccountIds:
          type: array
          items:
            type: string
          example: [ "acc_987654321", "acc_456789123" ]
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          example: "2025-08-18T10:15:30Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-08-18T10:15:30Z"

    # DEBIT CARD OPERATIONS
    AssociateAccountRequest:
      type: object
      properties:
        accountId:
          type: string
          example: "acc_789123456"
      required:
        - accountId
    DebitPurchaseRequest:
      type: object
      properties:
        amount:
          type: number
          format: double
          minimum: 0.01
          example: 150.00
        merchantInfo:
          type: string
          example: "Amazon Purchase"
        transactionType:
          type: string
          enum: [ "PURCHASE", "WITHDRAWAL", "PAYMENT" ]
          example: "PURCHASE"
      required:
        - amount
        - transactionType
    DebitPurchaseResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        transactionId:
          type: string
          example: "txn_123456789"
        cardId:
          type: string
          example: "64f3cbb5f8a1e53d2f7c34e2"
        requestedAmount:
          type: number
          format: double
          example: 150.00
        processedAmount:
          type: number
          format: double
          example: 150.00
        accountsUsed:
          type: array
          items:
            type: object
            properties:
              accountId:
                type: string
              amountDeducted:
                type: number
                format: double
              remainingBalance:
                type: number
                format: double
          example: [
            {
              "accountId": "acc_123456789",
              "amountDeducted": 100.00,
              "remainingBalance": 500.00
            },
            {
              "accountId": "acc_987654321",
              "amountDeducted": 50.00,
              "remainingBalance": 200.00
            }
          ]
        errorCode:
          type: string
          enum: [ "CARD_INACTIVE", "INSUFFICIENT_FUNDS", "INVALID_AMOUNT", "ACCOUNT_UNAVAILABLE" ]
        errorMessage:
          type: string
        processedAt:
          type: string
          format: date-time
    PrimaryAccountBalanceResponse:
      type: object
      required:
        - cardId
        - primaryAccountId
        - accountNumber
        - balance
        - currency
        - accountType
      properties:
        cardId:
          type: string
          example: "64f3cbb5f8a1e53d2f7c34e2"
        primaryAccountId:
          type: string
          example: "acc_123456789"
        accountNumber:
          type: string
          example: "0011-0123-456789"
        balance:
          type: number
          format: double
          example: 1500.50
        currency:
          type: string
          example: "PEN"
        accountType:
          type: string
          enum: [ "SAVINGS", "CHECKING" ]
          example: "SAVINGS"
        lastMovementDate:
          type: string
          format: date-time
          example: "2025-09-01T10:30:00Z"
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          example: "PERSON_ALREADY_HAS_CREDIT_CARD"
        message:
          type: string
          example: "Personal customers can only have one active credit card"
        timestamp:
          type: string
          format: date-time
          example: "2025-08-24T10:15:30Z"
      required:
        - code
        - message
        - timestamp